Provide native tests for Integrations resource
- it allows to have a non-regression test for #617 issue with
expressions
- extracted a single test from IntegrationsResourceTest which cannot be
launch in Native mode because it is using @Inject to access internal
services directly. it allows to inherit the native test class from
IntegrationsResourceTest and play all of the existing tests in native

part of #617
part of #437
