+++ api/src/test/java/io/kaoto/backend/api/resource/v1/IntegrationsResourceIT.java
@@ -1,0 +1,8 @@
+package io.kaoto.backend.api.resource.v1;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+class IntegrationsResourceIT extends IntegrationsResourceTest {
+
+}
